[#9] Rename: change file name from roadmap.ts to roadmapData.ts
constants의 roadmap.ts를 roadmapData.ts로 파일명 변경
